# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MojoPedia API",
        "description": "REST API untuk sistem manajemen pengetahuan MojoPedia. Menyediakan endpoint untuk manajemen user, role, dan integrasi Keycloak SSO.",
        "contact": {
            "name": "Tim Pengembang MojoPedia",
            "email": "admin@mojopedia.go.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/forum/threads": {
            "get": {
                "tags": [
                    "Forum"
                ],
                "summary": "Daftar thread diskusi forum (Publik)",
                "description": "Menampilkan seluruh diskusi yang sudah disetujui, mendukung filter kategori, sorting terpopuler/terbaru, dan pagination.",
                "operationId": "getForumThreads",
                "parameters": [
                    {
                        "name": "kategori",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "terbaru",
                                "terpopuler"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar thread forum",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Forum"
                ],
                "summary": "Buat topik diskusi forum baru (Sanctum)",
                "operationId": "createForumThread",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "judul",
                                    "konten",
                                    "category_id"
                                ],
                                "properties": {
                                    "judul": {
                                        "type": "string",
                                        "example": "Diskusi Implementasi SPBE 2026"
                                    },
                                    "konten": {
                                        "type": "string",
                                        "example": "Bagaimana evaluasi arsitektur domain layanan di OPD Anda?"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "knowledge_id": {
                                        "type": "integer",
                                        "example": 5,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Thread forum berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi form gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/forum/threads/{id}": {
            "get": {
                "tags": [
                    "Forum"
                ],
                "summary": "Detail thread diskusi beserta balasan / komentar",
                "operationId": "getForumThreadDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail thread dan balasan"
                    },
                    "404": {
                        "description": "Thread tidak ditemukan"
                    }
                }
            }
        },
        "/forum/threads/{id}/replies": {
            "post": {
                "tags": [
                    "Forum"
                ],
                "summary": "Kirim balasan / komentar pada thread (Sanctum)",
                "operationId": "replyForumThread",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "konten"
                                ],
                                "properties": {
                                    "konten": {
                                        "type": "string",
                                        "example": "Sangat setuju, kami sedang menyusun SOP integrasi."
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Balasan berhasil dikirim"
                    },
                    "400": {
                        "description": "Thread dikunci"
                    },
                    "404": {
                        "description": "Thread tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum Anda. Dapatkan token melalui endpoint login.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}